requirements.txt
Bio==1.8.0
biopython==1.81
click==8.1.7
dgl==2.2.1+cu118
fair_esm==2.0.0
matplotlib==3.6.2
numpy==2.3.3
pandas==1.5.0
scikit_learn==1.3.1
scipy==1.16.2
torch==2.4.1+cu118
tqdm==4.66.5
certifi==2023.7.22
charset-normalizer==3.3.0
cmake==3.27.7
fair-esm==2.0.0
filelock==3.12.4
fsspec==2023.9.2
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
lit==17.0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
packaging==23.2
Pillow==10.1.0
psutil==5.9.6
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
triton==2.0.0